requirements.txt
asgiref==3.8.1
bcrypt==4.1.3
cffi==1.16.0
cryptography==42.0.8
decorator==5.1.1
Deprecated==1.2.14
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fabric==3.2.2
invoke==2.2.0
paramiko==3.4.0
pip==24.0
pycparser==2.22
PyJWT==2.8.0
PyNaCl==1.5.0
setuptools==58.1.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
wheel==0.43.0
wrapt==1.16.0
